Fixes linter errors
