+++ torch/cuda/__init__.py
@@ -1282,8 +1282,6 @@ def _get_rng_state_offset(device: Union[int, str, torch.device] = "cuda") -> int
 
 from .memory import *  # noqa: F403
 
-from .mem_pool import *  # noqa: F403
-
 from .random import *  # noqa: F403
 
 ################################################################################
@@ -1624,7 +1622,6 @@ __all__ = [
     "memory_usage",
     "MemPool",
     "MemPoolContext",
-    "mem_pool",
     "temperature",
     "power_draw",
     "clock_rate",
